requirements.txt
streamlit  
langchain  
langchain-community  
openai  
faiss-cpu
pymupdf
httpx
pymilvus
sentence-transformers
watchdog